backend/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==4.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-cors-headers==3.10.1
django-extensions==3.1.5
django-extra-fields==3.0.2
django-heroku==0.0.0
django-imagekit==4.1.0
django-mysql==4.5.0
djangorestframework==3.13.1
drf-extra-fields==3.2.1
gunicorn==20.1.0
pilkit==2.0
Pillow==9.0.0
psycopg2==2.9.3
pydotplus==2.0.2
# pygraphviz==1.8
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==5.3.0
asgiref==3.4.1
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==36.0.1
daphne==3.0.2
dj-database-url==0.5.0
Django==4.0
django-appconf==1.0.5
django-autoslug==1.9.8
django-cors-headers==3.10.1
django-extensions==3.1.5
django-extra-fields==3.0.2
django-heroku==0.0.0
django-imagekit==4.1.0
django-mysql==4.5.0
djangorestframework==3.13.1
drf-extra-fields==3.2.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
pilkit==2.0
Pillow==9.0.0
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydotplus==2.0.2
# pygraphviz==1.8
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytz==2021.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.2.0
txaio==22.2.1
typing_extensions==4.1.1
whitenoise==5.3.0
zope.interface==5.4.0
